Changed project file structure
